Refactor integration tests
Makes everything a bit smoother
