+++ tests/cli.rs
@@ -1,37 +1,30 @@
-use assert_cmd::Command;
-
-#[path = "../src/test_helpers.rs"]
 mod test_helpers;
 use test_helpers::*;
 
 #[test]
-fn install_build_from_remote() {
-    logger();
+fn install_build_from_remote_directory() {
+    let (local, remote) = init();
+    let (local, remote) = (local.path(), remote.path());
 
-    let local = tempdir().unwrap();
-    let remote = tempdir().unwrap();
+    fs::write(remote.join("build1.tar.zst"), b"foobar").unwrap();
+    fs::write(remote.join("build2.tar.zst"), b"foobarbaz").unwrap();
 
-    fs::write(remote.path().join("build1.tar.zst"), b"foobar").unwrap();
-    fs::write(remote.path().join("build2.tar.zst"), b"foobarbaz").unwrap();
-
-    artefacta(local.path(), remote.path())
+    artefacta(local, remote)
         .args(&["install", "build2"])
         .assert()
         .success();
 
-    // Added "current" symlink
-    let current = local.path().join("current");
-    assert!(current.exists());
+    let current = local.join("current");
+    assert!(current.exists(), "Added `current` symlink");
 
-    // symlink points to new build that was also copied to local storage
-    let curent_path = fs::read_link(&current).unwrap();
-    assert_eq!(local.path().join("build2.tar.zst"), curent_path);
-}
+    assert!(
+        local.join("build2.tar.zst").exists(),
+        "new build was copied to local storage"
+    );
 
-fn artefacta(local: &Path, remote: &Path) -> Command {
-    let mut cmd = Command::cargo_bin("artefacta").unwrap();
-    cmd.env("ARTEFACTA_LOCAL_STORE", local);
-    cmd.env("ARTEFACTA_REMOTE_STORE", remote);
-    cmd.arg("--verbose");
-    cmd
+    assert_eq!(
+        local.join("build2.tar.zst"),
+        fs::read_link(&current).unwrap(),
+        "symlink points to new build"
+    );
 }
